Created enemyAI
Created the enemyAI along with IDamage script. Right now just have it so the enemy moves/reacts when player is near.
However, the enemy is not moving towards the player and I am not sure if that is because there are private variables in the game manager script that are set up for _playerState.