Multiple Factors for Enemy and UI
enemyAI has been modified so the spitter works at a basic level. i am working to upgrade it.

PlayerDamageScreen
PlayerPoisonScreen have been added to UI

Removed extra UI prefabs. If you are going to update the prefab make sure you save it to the UI Main, don't make a bunch of variants.

SpitterEnemy Prefab is set, the multi colored ones will be remade and adjusted but they are removed for now. Again, if you update the prefab make sure it saves to the Prefab folder and not just your Scene.