Labels Turn On/Off With Colliders
I have added a script and 2 colliders for turning the labels on/off. They are simple but will allow the player to not see the labels when they go inside the houses. Just make sure that each special label in the label folder have the tag "Label".
I have put them in prefabs but they should work now.