Lots of Updates
Have made the game functional from beginning to end for picking up objects and pathing through to the Pentagram. Along the way the player has a task list that will update with what to do next. Finally, the player can pick up the gun and escape the level.
Still need to implement the gun becoming available and useable, ranged enemies will cause damage but melee is broken. Working on that as well. Will place mines too to make the running away dynamic a little more and give us an additional enemy.